Stop deletion of location
Refactored Services to return bool values for location deletion.
Changed location deletion view to show message that location wasn't
deleted
Fixed issues with new (r)event overview pages

+++ EventCalendarBelle/EventCalendar/backoffice/ecTree/location.delete.controller.js
@@ -1,33 +1,27 @@
 ﻿angular.module("umbraco").controller("EventCalendar.LocationDeleteController",
         function ($scope, $routeParams, locationResource, notificationsService, navigationService, treeService) {
 
-            //$scope.delete = function (id) {
-            //    locationResource.deleteById(id).then(function (response) {
-            //        notificationsService.success("Success", $scope.currentNode.name + " has been deleted");
-            //        navigationService.reloadNode($scope.currentNode.parent);
-            //        navigationService.hideNavigation();
-            //    }, function (response) {
-            //        notificationsService.error("Error", $scope.currentNode.name + " could not be deleted");
-            //    });
-            //};
-
             $scope.performDelete = function () {
 
-                //mark it for deletion (used in the UI)
-                $scope.currentNode.loading = true;
-
-                locationResource.deleteById($scope.currentNode.id.replace("l-","")).then(function () {
+                locationResource.deleteById($scope.currentNode.id.replace("l-","")).then(function (response) {
                     $scope.currentNode.loading = false;
+                    if (response.data === true) {
+                        //mark it for deletion (used in the UI)
+                        $scope.currentNode.loading = true;
 
-                    //get the root node before we remove it
-                    var rootNode = treeService.getTreeRoot($scope.currentNode);
+                        //get the root node before we remove it
+                        var rootNode = treeService.getTreeRoot($scope.currentNode);
 
-                    //TODO: Need to sync tree, etc...
-                    treeService.removeNode($scope.currentNode);
+                        //TODO: Need to sync tree, etc...
+                        treeService.removeNode($scope.currentNode);
 
-                    navigationService.hideMenu();
+                        navigationService.hideMenu();
 
-                    notificationsService.success("Success", $scope.currentNode.name + " has been deleted");
+                        notificationsService.success("Success", $scope.currentNode.name + " has been deleted");
+                    } else {
+                        $scope.currentNode.loading = false;
+                        notificationsService.error("Error", $scope.currentNode.name + " could not be deleted");
+                    }
 
                 }, function () {
                     $scope.currentNode.loading = false;
